multi-node/aws: Setup internal LB for apiservers.
The apiserver loadbalancer is internal and attached to the Kubernetes subnet. An internal hostname for the LB is registered with Route 53.  It has the form "kubernetes.<cluster_name>.cluster.local".  This LB name is added to the apiserver TLS server certificate.  The worker kubelets are pointed to the LB name.