multi-node/aws: Set up etcd clustering among controller nodes.
Rather than use etcd's discovery mechanism or a static configuration, this makes use of the monsantoco/ectd-aws-cluster container, which makes use of AWS tools to discover the members of the controllers autoscale group.  This container is capable not only of starting a new cluster, but of managing removal and addition of cluster members as the membership of the autoscale group changes, for a fully automated solution.

References and access to etcd are removed from  worker nodes, as my branch no longer makes use of flannel, which was the only etcd client on the workers.